# Application/PlayerInfoService/api/apiPb/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PlayersService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/players": {
      "get": {
        "operationId": "PlayersService_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "PlayersService"
        ]
      },
      "delete": {
        "operationId": "PlayersService_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteRequest"
            }
          }
        ],
        "tags": [
          "PlayersService"
        ]
      },
      "post": {
        "operationId": "PlayersService_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AddResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddRequest"
            }
          }
        ],
        "tags": [
          "PlayersService"
        ]
      },
      "put": {
        "operationId": "PlayersService_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "tags": [
          "PlayersService"
        ]
      }
    }
  },
  "definitions": {
    "AddRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "club": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        }
      },
      "title": "---------------------------------------------------------------------------------------------------------------------\r\nAdd endpoint messages\r\n---------------------------------------------------------------------------------------------------------------------"
    },
    "AddResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "DeleteRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "---------------------------------------------------------------------------------------------------------------------\r\nDelete endpoint messages\r\n---------------------------------------------------------------------------------------------------------------------"
    },
    "DeleteResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "boolean"
        }
      }
    },
    "ListResponse": {
      "type": "object",
      "properties": {
        "players": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListResponsePlayer"
          }
        }
      }
    },
    "ListResponsePlayer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "club": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        }
      }
    },
    "UpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "club": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        }
      },
      "title": "---------------------------------------------------------------------------------------------------------------------\r\nUpdate endpoint messages\r\n---------------------------------------------------------------------------------------------------------------------"
    },
    "UpdateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}